requirements.txt
aiokafka==0.7.2
aioredis==2.0.0
alembic==1.7.5
autopep8==1.6.0
fastapi==0.70.0
fastapi-pagination[sqlmodel]==0.9.1
fastapi-utils==0.2.1
flake8==4.0.1
loguru==0.5.3
orjson==3.6.4
passlib[bcrypt]==1.7.4
psycopg2==2.9.2
pydantic[email]==1.8.2
PyJWT==2.3.0
pytest==6.2.5
pytest-cov==3.0.0
python-dotenv==0.19.1
python-multipart==0.0.5
requests==2.26.0
sqlmodel==0.0.4
uuid==1.30
uvicorn[standard]==0.15.0